revert refactor change
