+++ services/repository/fork.go
@@ -194,7 +194,7 @@ func ForkRepository(ctx context.Context, doer, owner *user_model.User, opts Fork
 		log.Error("Open created git repository failed: %v", err)
 	} else {
 		defer gitRepo.Close()
-		if err := SyncReleasesWithTags(ctx, repo, gitRepo); err != nil {
+		if err := repo_module.SyncReleasesWithTags(ctx, repo, gitRepo); err != nil {
 			log.Error("Sync releases from git tags failed: %v", err)
 		}
 	}
